Refactor database connections
