+++ index.js
@@ -5,8 +5,10 @@ require('dotenv').config();
 
 const PORT = process.env.PORT || 3030;
 
-openConnection().then(() => {
-  app.listen(PORT, () => {
-    logger.info(`Relate-With-Me is runing in ${process.env.NODE_ENV} mode on port ${PORT}.`);
-  });
-}).catch(err => logger.error(err));
+openConnection()
+  .then(() => {
+    app.listen(PORT, () => {
+      logger.info(`Relate-With-Me is runing in ${process.env.NODE_ENV} mode on port ${PORT}.`);
+    });
+  })
+  .catch(err => logger.error(err));
